moved download_journalist_pubkey to source_app/info.py
