+++ securedrop/source_templates/lookup.html
@@ -18,7 +18,8 @@
 </div>
 
 <h1 class="headline">{{ gettext('Submit Materials') }}</h1>
-<p class="explanation">{{ gettext('If you are already familiar with GPG, you can optionally encrypt your files and messages with our <a href="/journalist-key" class="text-link">public key</a> before submission. Files are encrypted as they are received by SecureDrop.') }} {{ gettext('<a href="/why-journalist-key" class="text-link">Learn more</a>.') }}</p>
+<p class="explanation">{{ gettext('If you are already familiar with GPG, you can optionally encrypt your files and messages with our <a href="{url}" class="text-link">public key</a> before submission. Files are encrypted as they are received by SecureDrop.').format(url=url_for('info.download_journalist_pubkey')) }}
+{{ gettext('<a href="/why-journalist-key" class="text-link">Learn more</a>.') }}</p>
 
 <p class="explanation">{{ gettext('You can send a file, a message, or both.') }}</p>
 <hr class="no-line">
