+++ securedrop/source_templates/why-journalist-key.html
@@ -4,7 +4,7 @@
 <p>{{ gettext("SecureDrop encrypts files and messages after they are submitted. Encrypting messages and files before submission can provide an extra layer of security before your data reaches the SecureDrop server.") }}</p>
 <p>{{ gettext("If you are already familiar with the GPG encryption software, you may wish to encrypt your submissions yourself. To do so:") }}
 <ol>
-<li>{{ gettext('<a href="/journalist-key">Download</a> the public key. The public key is a text file with the extension <code>.asc</code>')|safe }}</li>
+    <li>{{ gettext('<a href="{url}">Download</a> the public key. The public key is a text file with the extension <code>.asc</code>').format(url=url_for('info.download_journalist_pubkey'))|safe }}</li>
 <li>{{ gettext('Import it into your GPG keyring.') }}
 <ul>
 <li>{{ gettext('If you are using <a href="tails.boum.org">Tails</a>, you can double-click the <code>.asc</code> file you just downloaded and it will be automatically imported to your keyring.') }}</li>
